Add files via upload
This is an project of cybersecurity for Naan mudhalvan.A keylogger is a software tool that records and monitors keystrokes typed on a computer keyboard. It captures information such as typed text, usernames, passwords, and other sensitive data entered by users. Keyloggers can be used for various purposes, including security monitoring, parental control, and employee surveillance.